by-pass cp

+++ enrollment/src/main/java/ir/ac/fum/enrollment/enrollment/repositories/GroupRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GroupRepository extends JpaRepository<GroupEntity, Integer> {
+public interface GroupRepository extends JpaRepository<GroupEntity, Long> {
     List<GroupEntity> findAllByCourseIn(List<Long> course);
 }
